backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Cibi Aananth <cibiaananth@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
pydantic = {extras = ["mypy"], version = "^2.0.3"}
alembic = "^1.11.1"
sqlalchemy = {extras = ["asyncio", "mypy"], version = "^2.0.19"}
pydantic-settings = "^2.0.2"
passlib = "^1.7.4"
types-python-jose = "^3.3.4.8"
types-passlib = "^1.7.7.12"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
ruff = "^0.0.278"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
